Add files via upload
This MATLAB app has a simple user interface to take two piece of information, what the user wants to graph and for what country. The app then goes through large data sets to create a line graph with the required data. There are many COVID-19 tracking websites online, but this is different to those because it tracks how the outbreak started and how it developed. 

The purpose of this app is to be used to compare different countries to one another to see how they handled the corona virus pandemic and to use this information to find correlations and causations to other events. The next step will be to connect news articles from similar time periods to see how precautionary methods were implemented to prevent the spread of the virus. Also, a graph of the stock market such as the S&P 500 alongside to the outbreak in the relevant country could be useful to see how COVID-19 affected different things.

Overall, the app is designed to show the outbreak of COVID-19 in a simple manner to connect to other events that occurred worldwide to see how this pandemic affected the world.